requirements.txt
discord-ext-alternatives
discord
python-dotenv==0.14.0